% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFcsFile.R
\name{annotateFcsFile}
\alias{annotateFcsFile}
\title{Annotate FCS file}
\usage{
annotateFcsFile(experimentId, fcsFileId, annotations)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{fcsFileId}{ID of FCS file.}

\item{annotations}{List of annotations}
}
\description{
Sets the annotations on an FCS file.
}
\details{
This function overwrites any existing annotations. To add new annotations
while keeping the existing annotations, first use \code{getFcsFile} to get
the existing annotations. Append new annotations with \code{rbind}.

Note: if you retrieve FCS files in bulk, such as with \code{getFcsFiles},
\code{file$annotations} will return a nested list. Be sure to extract the
annotations from this list before appending new ones.
}
\examples{
\dontrun{
annotations = list(list(name="annotations 1", value=1), list(name="annotation 2", value="myValue"))
annotateFcsFile(experimentId, fcsFileId, annotations)
}
}
