% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyScaleSet.R
\name{applyScaleSet}
\alias{applyScaleSet}
\title{Apply a ScaleSet}
\usage{
applyScaleSet(scaleset, data, clamp_q = FALSE)
}
\arguments{
\item{scaleset}{ScaleSet. Accepts a ScaleSet or \code{scaleSet$scales}.}

\item{data}{Events data from an FcsFile, i.e. \code{events@expr}}

\item{clamp_q}{If true, values will be clamped to fall within the scale's
minimum and maximum. Applies clamping to all scales in a ScaleSet}
}
\description{
Applies a ScaleSet to a dataframe of events
}
\examples{
\dontrun{
getEvents(experimentId, fcsFileId, destination = "myFile.fcs")
events <- flowCore::read.FCS("myFile.fcs")
data <- events@exprs
scaleSet <- getScaleSets(experimentId)[1, ]
applyScaleSet(scaleSet, data)
}
}
