% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{Your username.}

\item{password}{Your password.}
}
\description{
Authenticates with the server. Call \code{\link{setServer}} before this
function.
}
\details{
For security reasons, you should not store your password in plain text, and
you should ensure that your R history does not store your password. Instead,
consider setting an environment variable and using something like
\code{authenticate("myusername", Sys.getenv("API_PASSWORD"))}
}
\examples{
\dontrun{
authenticate("username", "password")
}
}
