% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{byName}
\alias{byName}
\title{Resource By Name}
\usage{
byName(name)
}
\arguments{
\item{name}{Name of resource.}
}
\description{
Allows specifying a resource by name instead of by id, anywhere that an id is
accepted.
}
\details{
Internally, this looks up the resource's id by name before the function runs.
To improve performance, the resource's id is cached for the duration of the R
session. Resources such as gates that exist within an experiment are cached
within the experiment's scope. That is, the following is safe, even though
the FCS files have the same name:

\code{
getFcsFile(byName("experiment 1"), byName("fcsfile1.fcs"))
getFcsFile(byName("experiment 2"), byName("fcsfile1.fcs"))
}
}
\examples{
\dontrun{
getGates(byName("my experiment"))
getGates(experimentId = byName("my experiment"))
}
}
