% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFlowCore.R
\name{convertFromFlowCore}
\alias{convertFromFlowCore}
\title{Convert flowCore to gate}
\usage{
convertFromFlowCore(flowObject, scaleSet, name = NULL)
}
\arguments{
\item{flowObject}{A flowDensity object or a flowCore Gate}

\item{scaleSet}{The CellEngine ScaleSet to be applied.}
}
\description{
Converts points in a flowCore gate or flowDensity filter to a CellEngine
Gate and saves it to CellEngine. A ScaleSet must be provided to scale the
gate correctly.
}
\examples{
\dontrun{
sngl <- flowDensity(file, params)
scaleSet <- getScaleSets("5d2f8b4b21fd0676fb3a6a8c")
convertFromFlowCore(sngl, scaleSet, name = "my gate")
}
}
