% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFlowCore.R
\name{convertToFlowCore}
\alias{convertToFlowCore}
\title{Convert a gate to flowCore}
\usage{
convertToFlowCore(gate)
}
\arguments{
\item{gate}{The CellEngine gate to be converted}
}
\description{
Converts a CellEngine gate ("PolygonGate", "EllipseGate", or "RectangleGate") to
its flowCore analogue.
}
