% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarToParameters.R
\name{covarToParameters}
\alias{covarToParameters}
\title{Covert covariance matrix to ellipse parameters}
\usage{
covarToParameters(covar)
}
\arguments{
\item{covar}{Covariance matrix}
}
\description{
Returns list("major" = major, "minor" = minor, "angle" = angle)
}
