% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompensation.R
\name{createCompensation}
\alias{createCompensation}
\title{Create compensation}
\usage{
createCompensation(experimentId, name, spillMatrix, channels = NULL)
}
\arguments{
\item{experimentId}{The ID of the experiment of a \code{byName} expression.}

\item{name}{Name of the compensation matrix.}

\item{spillMatrix}{The spillover matrix. This can either be a 1-dimensional
vector of size \code{length(channels) ^ 2}, or a
\code{matrix}.}

\item{channels}{The channel names. Not required if \code{spillMatrix} is a
\code{matrix} with dimnames set.}
}
\description{
Creates a new compensation.
}
\examples{
\dontrun{
createCompensation(
  experimentId,
  name="Comp 1",
  channels=c("Ax488-A", "PE-A"),
  spillMatrix=c(1, 0,
                0, 1)
)
# same as:
createCompensation(
  experimentId,
  name="Comp 1",
  spillMatrix=matrix(
    c(1, 0,
      0, 1),
    nrow=2,
    ncol=2,
    byrow=TRUE,
    dimnames=list(c("Ax488-A", "PE-A"), c("Ax488-A", "PE-A"))
  )
)
}
}
