% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEllipseGate.R
\name{createEllipseGate}
\alias{createEllipseGate}
\title{Create ellipse gate}
\usage{
createEllipseGate(experimentId, xChannel, yChannel, name, x, y, angle,
  major, minor, label = c(x, y), gid = generateId(),
  parentPopulationId = NULL, parentPopulation = NULL,
  tailoredPerFile = FALSE, fcsFileId = NULL, fcsFile = NULL,
  locked = FALSE, createPopulation = TRUE)
}
\arguments{
\item{experimentId}{The ID of the experiment to which to add the gate.}

\item{xChannel}{The name of the x channel to which the gate applies.}

\item{yChannel}{The name of the y channel to which the gate applies.}

\item{name}{The name of the gate.}

\item{x}{The x centerpoint of the gate.}

\item{y}{The y centerpoint of the gate.}

\item{angle}{The angle of the ellipse in radians.}

\item{major}{The major radius of the ellipse.}

\item{minor}{The minor radius of the ellipse.}

\item{label}{Position of the label. Defaults to the midpoint of the gate.}

\item{gid}{Group ID of the gate, used for tailoring. If this is not specified,
then a new Group ID will be created.}

\item{parentPopulationId}{ID of the parent population. Use \code{NULL} for
the "ungated" population. If specified, do not specify \code{parentPopulation}.}

\item{parentPopulation}{Name of the parent population. An attempt will be made
to find the population by name. If zero or more than one population exists
with the name, an error will be thrown. If specified, do not specify
\code{parentPopulationId}.}

\item{tailoredPerFile}{Whether or not this gate is tailored per FCS file.}

\item{fcsFileId}{ID of FCS file, if tailored per file. Use \code{NULL} for
the global gate in a tailored gate group. If specified, do not specify
\code{fcsFile}.}

\item{fcsFile}{Name of FCS file, if tailored per file. An attempt will be made
to find the file by name. If zero or more than one file exists with the name,
an error will be thrown. If specified, do not specify \code{fcsFileId}.}

\item{locked}{Prevents modification of the gate via the web interface.}

\item{createPopulation}{Automatically create corresponding population.}
}
\description{
Creates an ellipse gate.
}
\examples{
\dontrun{
createEllipseGate(experimentId, "FSC-A", "FSC-W", "my gate", c(1, 2, 3), c(4, 5, 6))
}
}
