% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperiment.R
\name{createExperiment}
\alias{createExperiment}
\title{Create experiment}
\usage{
createExperiment(properties = list(), params = list())
}
\arguments{
\item{properties}{Optional properties for the new experiment.}

\item{params}{Optional query parameters.}
}
\description{
Creates a new experiment.
}
\examples{
\dontrun{
createExperiment() # creates a blank experiment
createExperiment(list("name" = "my experiment"))
}
}
