% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGates.R
\name{createGates}
\alias{createGates}
\title{Create gates in bulk.}
\usage{
createGates(experimentId, gates)
}
\arguments{
\item{experimentId}{The ID of the experiment to which to add the gate.}

\item{gates}{\code{list()} of gates.}
}
\description{
Creates multiple gates in a single API call. This is faster than calling
\code{createXxxGate} in a loop.
}
\details{
Each gate must be specified as a \code{list()} object conforming to our API
specification (TODO INSERT LINK WHEN PUBLIC); however, some properties will
have defaults automatically set if not specified (\code{model.locked},
\code{model.label}, \code{gid}, \code{parentPopulationId},
\code{tailoredPerFile}, \code{fcsFileId}, \code{experimentId}).

Up to ~1,000 gates may be created at once. If any single gate fails
validation on the server, no gates will be created.

Unlike \code{createXxxGate}, this function cannot automatically create
populations corresponding to each gate. This function is primarily suited to
tailoring a large number of gates.
}
\examples{
\dontrun{
g1 = list(type = "RectangleGate", xChannel = "FSC-A", yChannel = "SSC-A",
  model = list(rectangle = list(x1 = 1, x2 = 100, y1 = 1, y2 = 100)))
g2 = list(type = "PolygonGate", xChannel = "FSC-A", yChannel = "SSC-A",
  model = list(polygon = list(vertices = c(c(1, 2), c(30, 40), c(50, 60)))))
g3 = list(type = "RangeGate", xChannel = "V450-480-A",
  model = list(range = list(x1 = 1, x2 = 100, y = 0.5)))

createGates(experimentId, c(g1, g2, g3))
}
}
