% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{createLookup}
\alias{createLookup}
\title{Find a Resource By Name}
\usage{
createLookup(experimentId)
}
\arguments{
\item{experimentId}{Experiment to create lookup for.}
}
\value{
A function that takes the parameters "resource" and "name":
  resource Type of resource. Available options are "gates", "populations", "fcsfiles", and "compensations".
  name Name of resource to search for.
}
\description{
Allows creation of a by-name lookup object per experiment, which can
then be used to find resources by name instead of id. The resource's id is
cached for the duration of the R session.
If a name is not specified, a list of all requested resources will be returned.
}
\examples{
\dontrun{
lookup = createLookup("5e1f66a06f5f3f0759b479c9")
lookup("gates", "test_gate")
# or:
lookup("gates")
}
}
