% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPopulation.R
\name{createPopulation}
\alias{createPopulation}
\title{Create population}
\usage{
createPopulation(experimentId, name, gates, terminalGateGid, parentId = NULL)
}
\arguments{
\item{experimentId}{The ID of the experiment to which to add the population.}

\item{name}{The name of the population.}

\item{gates}{The list of gates comprising the population. This should be in
the form \code{list(`$and` = c("gid1", "gid2"))}, where \code{"gid1"} and
\code{"gid2"} are the Group IDs (\code{gid}s) of the gates. Any combination
of \code{$and}, \code{$or}, \code{$not} and \code{$xor} can be used;
however, only populations with the form \code{list(`$and` = c(...gates))}
can be displayed in the Web UI at this time.}

\item{terminalGateGid}{The Group ID (\code{gid}) of the gate that
differentiates this population from its parent. Because complex populations
that deviate from the \code{list(`$and` = c(...gates))} form cannot currently
be displayed in the Web UI, those populations may set this to \code{NULL}.
In the future, this will be required to be a list of GIDs for those
populations.}

\item{parentId}{ID of the parent population. Use \code{NULL} for the
"ungated" population.}
}
\description{
Creates a population.
}
\examples{
\dontrun{
gid1 = "59262d84b1a1fc1193f12b0e"
createPopulation(experimentId, "Singlets", list(`$and` = c(gid1)), gid1)
}
}
