% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSplitGate.R
\name{createSplitGate}
\alias{createSplitGate}
\title{Create split gate}
\usage{
createSplitGate(
  experimentId,
  xChannel,
  name,
  x,
  y = 0.5,
  gid = generateId(),
  gids = replicate(2, generateId()),
  labels = NULL,
  parentPopulationId = NULL,
  tailoredPerFile = FALSE,
  fcsFileId = NULL,
  locked = FALSE,
  createPopulations = is.null(fcsFileId)
)
}
\arguments{
\item{experimentId}{The ID of the experiment to which to add the gate, or a
\code{byName} expression.}

\item{xChannel}{The name of the x channel to which the gate applies.}

\item{name}{The name of the gate. Each sector is named with this parameter
and a split flag (L, R).}

\item{x}{The x coordinate of the center point (after the channel's scale has been applied).}

\item{y}{The y position of the dashed line extending from the center point (as a fraction
of the height).}

\item{gid}{Top-level group ID of the gate, used for tailoring. If this is not specified,
then a new Group ID will be created. For compound gates, "gid" refers to the
top-level GID. Each sector has a unique model gid and name to which
populations must refer.}

\item{gids}{Group IDs of each sector, assigned to model.gids.}

\item{labels}{Positions of the quadrant labels. A list of two length-2 vectors in
the order: L, R. These are set automatically to the top corners.}

\item{parentPopulationId}{ID of the parent population or a \code{byName}
expression. Use \code{UNGATED} for the "ungated" population. Only used if
\code{createPopulation} is \code{TRUE}.}

\item{tailoredPerFile}{Whether or not this gate is tailored per FCS file.}

\item{fcsFileId}{ID of FCS file or a \code{byName} expression, if tailored
per file. Use \code{NULL} for the global gate in a tailored gate group.}

\item{locked}{Prevents modification of the gate via the web interface.}

\item{createPopulations}{Automatically create corresponding populations.
Specify \code{parentPopulationId} if set to \code{TRUE}. Because R performs
partial matching of argument names, this can also be specified as
\code{createPopulation} for consistency with the other \code{create__Gate}
functions.}
}
\description{
Creates a split gate. split gates have two sectors (right and left),
each with a name and unique gid.
}
\examples{
\dontrun{
createSplitGate(experimentId, "FSC-A", "my gate", 144000, 100000)
}
}
