% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCompensation.R
\name{deleteCompensation}
\alias{deleteCompensation}
\title{Delete compensation}
\usage{
deleteCompensation(experimentId, compensationId)
}
\arguments{
\item{experimentId}{ID of experiment or a \code{byName} expression.}

\item{compensationId}{ID of compensation or a \code{byName} expression.}
}
\description{
Deletes a compensation.
}
\examples{
\dontrun{
deleteCompensation(experimentId, compensationId)
}
}
