% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteExperiment.R
\name{deleteExperiment}
\alias{deleteExperiment}
\title{Delete experiment}
\usage{
deleteExperiment(experimentId)
}
\arguments{
\item{experimentId}{ID of experiment.}
}
\description{
Deletes an experiment.
}
\examples{
\dontrun{
deleteExperiment(experimentId)
}
}
