% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteFcsFile.R
\name{deleteFcsFile}
\alias{deleteFcsFile}
\title{Delete FCS file}
\usage{
deleteFcsFile(experimentId, fcsFileId)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{fcsFileId}{ID of FCS file.}
}
\description{
Deletes an FCS file.
}
\examples{
\dontrun{
deleteFcsFile(experimentId, fcsFileId)
}
}
