% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteGates.R
\name{deleteGates}
\alias{deleteGates}
\title{Delete gate}
\usage{
deleteGates(experimentId, gid = NULL, gateId = NULL, exclude = NULL)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{gid}{ID of gate family.}

\item{gateId}{ID of gate.}

\item{exclude}{Gate ID to exclude from deletion.}
}
\description{
Deletes a gate or a tailored gate family
}
\details{
This is an advanced function, primarily used for deleting tailored gates. It
does not delete populations. Use \code{deletePopulationAuto} to delete gates
and their associated populations.

Works for compound gates if you specify the top-level gid. Specifying the gid
of a sector (i.e. one listed in model.gids) will result in no gates being
deleted. If gateId is specified, only that gate will be deleted, regardless
of the other parameters specified.

Currently, \code{byName} does not work for the gateId in this function (it does work
for the experimentId).
}
\examples{
\dontrun{
deleteGate(experimentId, gid = gateFamilyID)
deleteGate(experimentId, gateId = gateID)
}
}
