% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePopulationAuto.R
\name{deletePopulationAuto}
\alias{deletePopulationAuto}
\title{Auto Delete Population}
\usage{
deletePopulationAuto(experimentId, populationId)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{populationId}{ID of population.}
}
\description{
Deletes a population, its corresponding gates, and all descendant gates and
populations.
}
\examples{
\dontrun{
deletePopulationAuto(experimentId, populationId)
}
}
