% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAttachment.R
\name{downloadAttachment}
\alias{downloadAttachment}
\title{Download Attachment}
\usage{
downloadAttachment(
  experimentId,
  attachmentId,
  destination = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{experimentId}{ID of the experiment.}

\item{attachmentId}{ID of the attachment.}

\item{destination}{Optional. If specified, write the file to the specified
path instead of returning it as a binary blob.}

\item{overwrite}{Optional. If a destination is specified, allows destination
file to be overwritten.}
}
\description{
Downloads an attachment and returns it as a binary blob or saves it to disk.
}
\examples{
\dontrun{
# Returns the attachment as a binary blob:
attachment = downloadAttachment(experimentId, attachmentId)
# Parse it as text:
readBin(attachment, character())

# Saves the attachment to disk:
downloadAttachment(experimentId, attachmentId, destination = "dir/file.txt")

# Use the byName helper to avoid finding the attachment ID:
downloadAttachment(experimentId, byName("file.txt"))
}
}
