% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFcsFiles.R
\name{downloadFcsFiles}
\alias{downloadFcsFiles}
\title{Download multiple FCS files}
\usage{
downloadFcsFiles(
  experimentId,
  format,
  destination,
  overwrite = F,
  fcsFileIds = NULL,
  populationIds = NULL,
  compensationId = NULL,
  compensatedQ = NULL,
  preSubsampleN = NULL,
  preSubsampleP = NULL,
  postSubsampleN = NULL,
  postSubsampleP = NULL,
  seed = NULL,
  filenameTemplate = NULL
)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{format}{Specifies the file format ("fcs", "tsv (with header)" or
"tsv (without header)")}

\item{destination}{[String] Write the files to the specified destination.}

\item{overwrite}{[Logical] Allows a destination file to be overwritten.}

\item{fcsFileIds}{[Array<String>] Optional. Defaults to returning all files
in the experiment.}

\item{populationIds}{[Array<String>] Optional. If provided, only events from
these populations will be included in the output files. Defaults to ungated.}

\item{compensationId}{[String] Required if populationIds is specified.
Compensation to use for gating.}

\item{compensatedQ}{[Logical] Optional. If true, applies the compensation
specified in compensationId to the exported events. For TSV format, the
numerical values will be the compensated values. For FCS format, the
numerical values will be unchanged, but the file header will contain the
compensation as the spill string (file-internal compensation).}

\item{preSubsampleN}{[Integer] Randomly subsample the file to contain this many events
before gating.}

\item{preSubsampleP}{[Numeric] Randomly subsample the file to contain this percent of
events (0 to 1) before gating.}

\item{seed}{[Numeric] Seed for random number generator used for subsampling. Use for
deterministic (reproducible) subsampling. If omitted, a pseudo-random value
is used.}

\item{filenameTemplate}{[String]    Tokenized template to dynamically name each
file in the resulting archive, like "{file.name}-{population.name}"}

\item{postSubSampleN}{[Integer] Randomly subsample the file to contain this many events
after gating.}

\item{postSubSampleP}{[Numeric] Randomly subsample the file to contain this percent of
events (0 to 1) after gating.}
}
\description{
Downloads multiple files bundled into a ZIP archive.
}
\examples{
\dontrun{
# Download all FCS files in the experiment
downloadFcsFiles(experimentId, "fcs", "archive.zip", overwrite = T)
# Download specific FCS file
downloadFcsFiles(experimentId, "fcs", "archive.zip", fcsFileIds=c("5d2f8b4b21fd0676fb3a6a72", "5d2f8b4b21fd0676fb3a6a74"))
}
}
