% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{FILE_INTERNAL}
\alias{FILE_INTERNAL}
\title{File-internal compensation.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
FILE_INTERNAL
}
\description{
A constant representing file-internal compensation.
}
\keyword{datasets}
