% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEllipsePoints.R
\name{fitEllipsePoints}
\alias{fitEllipsePoints}
\title{Fit ellipse points}
\usage{
fitEllipsePoints(xy, tolerance = 0.005, max.iter = 500)
}
\arguments{
\item{xy}{a two-column data frame containing x and y coordinates.}

\item{tolerance}{a tolerance value (default = 0.005).}

\item{max.iter}{Maximum algorithm iterations.}
}
\value{
A named list containing:
 - "covar", the 2x2 covariance matrix of the ellipse
 - "center", a 2x1 vector of the center of the ellipse such that:
     (x - C)' A (x - C) <= 1
 - "axes", a 2x1 vector whose elements are one-half the
     lengths of the major and minor axes
 - "angle", the angle of rotation
}
\description{
Calculate the minimum volume enclosing ellipsoid for a set of points
}
\details{
Uses the Khachiyan Algorithm to find the the minimum volume enclosing
ellipsoid (MVEE) of a set of points. In two dimensions, this is just
the bounding ellipse (this function is limited to two dimensions).
Adapted from https://stat.ethz.ch/pipermail/r-help/2011-March/272997.html
}
