% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFlowCore.R
\name{fromFlowCore}
\alias{fromFlowCore}
\title{Convert a flowCore object to a CellEngine object}
\usage{
fromFlowCore(flowCoreObj, experimentId, name = NULL, ...)
}
\arguments{
\item{flowCoreObj}{A flowCore object.}

\item{experimentId}{The ID of the experiment in which to create the object.}

\item{name}{The name to use for the CellEngine object. If not set, it will
default to the identifier of the flowCore object.}

\item{...}{Other parameters passed to \code{createRectangleGate},
\code{createPolygonGate}, etc.}
}
\description{
Creates a CellEngine object from a flowCore object, saving it to CellEngine.
}
\examples{
\dontrun{
# Converting a flowCore rectangleGate:
rectGate <- rectangleGate(
  filterId = "Gate 1",
  "FL1-H" = c(0, 12), "FL2-H" = c(0, 12)
)
experimentId <- "5d2f8b4b21fd0676fb3a6a8c"
fromFlowCore(rectGate, experimentId, name = "my gate")
# "my gate" should now exist in the CellEngine experiment.
}
}
