% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateId.R
\name{generateId}
\alias{generateId}
\title{Generate ID}
\usage{
generateId()
}
\value{
an ID string
}
\description{
Generates an ID string suitable for use as a \code{gid} when creating gates.
}
\examples{
generateId()
}
