% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttachments.R
\name{getAttachments}
\alias{getAttachments}
\title{Get experiment attachments}
\usage{
getAttachments(experimentId, params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{params}{Optional query parameters.}
}
\description{
Retrieves the list of attachments in an experiment.
}
\examples{
\dontrun{
getAttachments(experimentId)
getAttachments(experimentId, params = list("limit" = "5"))
}
}
