% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompensation.R
\name{getCompensation}
\alias{getCompensation}
\title{Get compensation}
\usage{
getCompensation(experimentId, compensationId)
}
\arguments{
\item{experimentId}{ID of experiment or a \code{byName} expression.}

\item{compensationId}{ID of compensation or a \code{byName} expression.}
}
\description{
Retrieves a compensation. The returned object has a \code{spillMatrix}
property that is a \code{matrix}.
}
\examples{
\dontrun{
# Retrieve by ID
getCompensation(experimentId, compensationId)

# Lookup by name
getCompensation(experimentId, byName("Comp 1"))
}
}
