% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEllipsePoints.R
\name{getEllipsePoints}
\alias{getEllipsePoints}
\title{Get ellipse points}
\usage{
getEllipsePoints(x, y, angle, major, minor)
}
\description{
Returns 8 perimeter points at angular increments of pi / 4.
}
