% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getErrorInfo}
\alias{getErrorInfo}
\title{Returns more information about the last error.}
\usage{
getErrorInfo()
}
\description{
Some functions in this package will store additional error information so
that it is retrievable using \code{getErrorInfo()}.
}
