% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{Get FCS File Events}
\usage{
getEvents(experimentId, fcsFileId, populationId = NULL, scaleSetId = NULL,
  compensatedQ = FALSE, headerQ = FALSE, format = "FCS",
  destination = NULL)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{fcsFileId}{ID of FCS file.}

\item{populationId}{Optional ID of population, for gated events.}

\item{scaleSetId}{Optional ID of scale set. If omitted and if a single
scale set exists in the experiment, that scale set will be used.}

\item{compensatedQ}{For TSV format only: if true, file will contain
compensated values.}

\item{headerQ}{for TSV format only: if true, file will contain a header row.}

\item{format}{One of "TSV" or "FCS".}

\item{destination}{Optional, if specified, write the file to the specified
path instead of returning it as a binary blob.}
}
\description{
Retrieves events from an FCS file or gated population as either an FCS or
TSV file.
}
\details{
Tip: FCS files are more compact than TSV files. Use that format for faster
downloads. Use a library such as flowCore to parse FCS files.
}
\examples{
\dontrun{
getEvents(experimentId, fcsFileId)
getEvents(experimentId, fcsFileId, populationId, format = "TSV")
getEvents(experimentId, fcsFileId, destination = "/path/to/output.fcs")
}
}
