% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiment.R
\name{getExperiment}
\alias{getExperiment}
\title{Get experiment}
\usage{
getExperiment(experimentId, params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{params}{Optional query parameters.}
}
\description{
Retrieves an experiment.
}
\examples{
\dontrun{
getExperiment(experimentId)
getExperiment(experimentId, params = list("fields" = "+name"))
}
}
