% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{getExperiments}
\title{Get experiments}
\usage{
getExperiments(params = list())
}
\arguments{
\item{params}{Optional query parameters.}
}
\description{
Retrieves the list of experiments to which you have access.
}
\examples{
\dontrun{
# List all accessible experiments
getExperiments()

# List the names of the first five experiments
getExperiments(params = list("limit" = "5", "fields" = "+name"))
}
}
