% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFcsFile.R
\name{getFcsFile}
\alias{getFcsFile}
\title{Get FCS File}
\usage{
getFcsFile(experimentId, fcsFileId, params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{fcsFileId}{ID of FCS file.}

\item{params}{Optional query parameters.}
}
\description{
Retrieves an FCS file. This retrieves metadata about the file, but does not
retrieve the file's contents.
}
\examples{
\dontrun{
# Retrieve by ID
getFcsFile(experimentId, fcsFileId)

# Lookup by name
getFcsFile(experimentId, byName("Sample 1.fcs"))
}
}
