% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGate.R
\name{getGate}
\alias{getGate}
\title{Get a gate}
\usage{
getGate(experimentId, gateId)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{gateId}{ID of gate.}
}
\description{
Retrieves a gate from an experiment.
}
\examples{
\dontrun{
getGate(experimentId, gateId)
}
}
