% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGates.R
\name{getGates}
\alias{getGates}
\title{Get gates}
\usage{
getGates(experimentId, params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{params}{Optional query parameters.}
}
\description{
Retrieves the list of gates in an experiment.
}
\examples{
\dontrun{
getGates(experimentId)
getGates(experimentId, params = list("limit" = "5"))
}
}
