% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlot.R
\name{getPlot}
\alias{getPlot}
\title{Get a plot}
\usage{
getPlot(
  experimentId,
  fcsFileId,
  plotType,
  xChannel,
  yChannel = "",
  zChannel = "",
  params = list(),
  destination = NULL,
  overwrite = FALSE,
  display = TRUE
)
}
\arguments{
\item{experimentId}{character ID of experiment.}

\item{fcsFileId}{character ID of FcsFile}

\item{plotType}{character Plot type, one of: "contour", "dot", "density", or "histogram"}

\item{xChannel}{character X channel name}

\item{yChannel}{character Y channel name}

\item{zChannel}{character Z channel name}

\item{params}{list Optional query parameters:
\itemize{
\item axesQ logical Display axes lines. Defaults to \code{TRUE}.
\item axisLabelsQ logical Display axis labels. Defaults to \code{TRUE}.
\item color character Case-insensitive string in the format #rgb, #rgba,
#rrggbb or #rrggbbaa. The foreground color, i.e. color of curves in
"histogram" and "contour" plots, dots in "dot" plots.
\item compensation character ID of compensation to use for gating and
display, or one of \code{\link{UNCOMPENSATED}} or \code{\link{FILE_INTERNAL}}.
\item gateLabel character One of "eventcount", "mean", "median", "percent",
"mad", "cv", "stddev", "geometricmean", "name", "none".
\item gateLabelFontSize double Font size for gate labels.
\item height double Image height. Defaults to 228.
\item percentileStart double For contour plots, the percentile of the first
contour.
\item percentileStep double For contour plots, the percentile step between
each contour.
\item populationId character Population ID. Defaults to ungated.
\item postSubsampleN double Randomly subsample the file to contain this many
events after gating.
\item postSubsampleP double Randomly subsample the file to contain this
percent of events (0 to 1) after gating.
\item renderGates logical Render gates into the image.
\item seed double Seed for random number generator used for subsampling. Use
for deterministic reproducible subsampling. If omitted, a pseudo-random
value is used.
\item smoothing double For density and contour plots, adjusts the amount of
smoothing. Defaults to 0 (no smoothing). Set to 1 for typical smoothing.
Higher values (up to 10) increase smoothing.
\item strokeThickness double The thickness of histogram and contour plot
lines. Defaults to 1.
\item tickLabelsQ logical Display tick labels. Defaults to \code{FALSE}.
\item ticksQ logical Display ticks. Defaults to \code{TRUE}.
\item width Integer Image width. Defaults to 228.
\item xAxisLabelQ logical Display x axis label. Overrides axisLabelsQ.
\item xAxisQ logical Display x axis line. Overrides axesQ.
\item xTickLabelsQ logical Display x tick labels. Overrides tickLabelsQ.
\item xTicksQ logical Display x ticks. Overrides ticksQ.
\item yAxisLabel logical Display y axis label. Overrides axisLabelsQ.
\item yAxisQ logical Display y axis line. Overrides axesQ.
\item yTickLabelsQ logical Display y tick labels. Overrides tickLabelsQ.
\item yTicksQ logical Display y ticks. Overrides ticksQ.
}}

\item{destination}{character If specified, write the image to file.}

\item{overwrite}{logical If a destination is specified, allows destination
file to be overwritten.}

\item{display}{logical Display the image. Defaults to \code{TRUE}.}
}
\description{
Gets a plot for an FcsFile. If in RStudio or Jupyter and \code{display} is \code{TRUE}
(default), displays the image. If \code{destination} is provided, also saves the
image as a PNG file. In all cases, returns the PNG-encoded bytes (invisibly,
if \code{display} is \code{TRUE}).
}
\examples{
\dontrun{
parameters <- list(
  "width" = 400,
  "height" = 400,
  "axesQ" = FALSE,
  "ticksQ" = FALSE,
  "color" = "#ff0000",
  "renderGates" = TRUE,
  "smoothing" = 0.75,
)
getPlot(experimentId,
  fcsFileId,
  plotType = "dot",
  xChannel = "FSC-A",
  yChannel = "FSC-W",
  params = parameters
)
}
}
