% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaleSets.R
\name{getScaleSets}
\alias{getScaleSets}
\title{Get scale sets}
\usage{
getScaleSets(experimentId, params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{params}{Optional query parameters.}
}
\description{
Retrieves the list of scale sets in an experiment.
}
\examples{
\dontrun{
getScaleSets(experimentId)
getScaleSets(experimentId, params = list("limit" = "5"))
}
}
