% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{PER_FILE}
\alias{PER_FILE}
\title{Per-file compensation.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PER_FILE
}
\description{
A constant representing per-file compensation (a possible value for an
experiment's \code{activeCompensation}).
}
\keyword{datasets}
