% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFlowCore.R
\name{scaleFlowCoreGate}
\alias{scaleFlowCoreGate}
\title{Scale flowCore gate}
\usage{
scaleFlowCoreGate(gate, scaleSet)
}
\arguments{
\item{gate}{The flowCore gate points (i.e. PologonalGate@boundaries)}

\item{scaleSet}{The CellEngine ScaleSet.}
}
\description{
Applies a scale to points from a flowCore gate, so that it may
be saved to CellEngine.
}
\examples{
\dontrun{
sngl <- flowDensity(file, params)
flowGate <- sngl@filter
scaleSet <- getScaleSets("5d2f8b4b21fd0676fb3a6a8c")
gatePoints <- scaleFlowCoreGate(flowGate, scaleSet)
createPolygonGate(id, colnames(flowGate)[1], colnames(flowGate)[2], 'my gate', vertices=gatePoints)
}
}
