% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToFlowCore.R
\name{scaleSetToTransformList}
\alias{scaleSetToTransformList}
\title{Convert ScaleSet}
\usage{
scaleSetToTransformList(scaleSet)
}
\arguments{
\item{scaleSet}{The CellEngine scaleSet to be converted}
}
\description{
Convert a CellEngine ScaleSet to a flowCore transformList
}
