% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFcsFilePanel.R
\name{setFcsFilePanel}
\alias{setFcsFilePanel}
\title{Set FCS file panel}
\usage{
setFcsFilePanel(experimentId, fcsFileId, panelName, panel)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{fcsFileId}{ID of FCS file.}

\item{panelName}{Name of panel.}

\item{panel}{List of channel properties.}
}
\description{
Updates the panel properties of an FCS file.
}
\details{
A panel must be specified as a list of lists. Each inner list must have at
least \code{channel} and \code{index} properties, and may optionally have a
\code{reagent} property.

The \code{index} value corresponds to the parameter index (i.e. the n in
"$PnN" from the FCS file header). This property allows the list of channels
to be in any order.

The \code{channel} value corresponds to the channel name. Channel names are
used for matching channels between FCS files, gates, scales and compensations.

The \code{reagent} value is used for various displays, including plot axis
labels, and is typically the reagent description (e.g. "CD3-PacBlu").

The \code{panelName} property is used to group files by panel.
}
\examples{
\dontrun{
panel = list(
  list("index" = 1, "channel" = "FSC-A"),
  list("index" = 7, "channel" = "Blue530-A", "reagent" = "CD3")
)
setFcsFilePanel(experimentId, fcsFileId, "Panel 1", panel)
}
}
