% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setServer.R
\name{setServer}
\alias{setServer}
\title{Set server address}
\usage{
setServer(host)
}
\arguments{
\item{host}{Address of API server.}
}
\description{
Sets the address of the server to which to connect.
}
\examples{
setServer("https://mycompany.immuneatlas.org")
}
