% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toFlowCore.R
\name{toFlowCore}
\alias{toFlowCore}
\title{Convert a CellEngine object to flowCore}
\usage{
toFlowCore(cellengineObject)
}
\arguments{
\item{cellengineObject}{The CellEngine object to be converted.}
}
\description{
Converts a CellEngine object to its flowCore analogue.
}
\details{
Notes: Split gates and quadrant gates will be converted to lists of two
and four rectangle gates, respectively. Skewed quadrant gates from CellEngine
will be converted to orthogonal quadrant gates, as flowCore's quadGate class
does not support skewing.
}
\examples{
\dontrun{
# Converting a compensation:
ceComp <- getCompensation(experimentId, byName("My comp"))
fcComp <- toFlowCore(ceComp)
}
}
