% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{UNCOMPENSATED}
\alias{UNCOMPENSATED}
\title{No compensation.}
\format{An object of class \code{numeric} of length 1.}
\usage{
UNCOMPENSATED
}
\description{
A constant representing no compensation.
}
\keyword{datasets}
