% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCompensation.R
\name{updateCompensation}
\alias{updateCompensation}
\title{Update compensation}
\usage{
updateCompensation(experimentId, compensationId, properties = list())
}
\arguments{
\item{experimentId}{ID of experiment or a \code{byName} expression.}

\item{compensationId}{ID of compensation or a \code{byName} expression.}

\item{properties}{Properties to set on the compensation.}
}
\description{
Updates a compensation.
}
\examples{
\dontrun{
updateCompensation(experimentId, compensationId, list("name" = "new name"))
}

}
