% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExperiment.R
\name{updateExperiment}
\alias{updateExperiment}
\title{Update experiment}
\usage{
updateExperiment(experimentId, properties = list(), params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{properties}{Properties to set on the experiment.}

\item{params}{Optional query parameters.}
}
\description{
Updates an experiment.
}
\examples{
\dontrun{
updateExperiment(experimentId, list("name" = "my experiment"))
}
}
