% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFcsFile.R
\name{updateFcsFile}
\alias{updateFcsFile}
\title{Update Fcs File}
\usage{
updateFcsFile(experimentId, fcsFileId, properties = list())
}
\arguments{
\item{experimentId}{character ID of experiment.}

\item{fcsFileId}{character ID of file or a \code{byName} expression.}

\item{properties}{list Properties to set on the fcs file.

Accepted properties:
\itemize{
\item filename character The name of the file
\item panel list List with properties:
\itemize{
\item channel character
\item index character
\item reagent character Optional; only channel and index are required.
}
\item panelName character
\item isControl logical
\item compensation character Compensation ID
\item deleted string ISO 8601 date string, "\%Y-\%m-\%dT\%H:\%M:\%S.\%fZ"
\item gatesLocked logical
\item annotations list List with properties:
\itemize{
\item name character
\item value character
}
}}
}
\description{
Updates an fcs file
}
\examples{
\dontrun{
updateFcsFile(experimentId, list("filename" = "my fcs file"))
}
}
