% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGate.R
\name{updateGate}
\alias{updateGate}
\title{Update gate}
\usage{
updateGate(experimentId, gateId, properties = list(), params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{gateId}{Gate family ID}

\item{properties}{Properties to set on the experiment.}

\item{params}{Optional query parameters.}
}
\description{
Updates a gate.
}
\examples{
\dontrun{
updateGate(experimentId, gateId, list("name" = "new gate name"))
}
}
