% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGateFamily.R
\name{updateGateFamily}
\alias{updateGateFamily}
\title{Update gate family}
\usage{
updateGateFamily(experimentId, gid, properties = list(), params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{gid}{Gate family ID}

\item{properties}{Properties to set on the gate.}

\item{params}{Optional query parameters.}
}
\description{
Updates a gate family.
}
\examples{
\dontrun{
updateGateFamily(experimentId, gid, list("name" = "new gate name"))
}
}
