% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePopulation.R
\name{updatePopulation}
\alias{updatePopulation}
\title{Update population}
\usage{
updatePopulation(experimentId, populationId, properties = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{populationId}{Population ID or a \code{byName()} expression.}

\item{properties}{Properties to set on the population.}
}
\description{
Updates a population.
}
\examples{
\dontrun{
updatePopulation(experimentId, populationId, list("name" = "new pop name"))
}
}
