% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadAttachment.R
\name{uploadAttachment}
\alias{uploadAttachment}
\title{Upload attachment}
\usage{
uploadAttachment(experimentId, attachmentPath)
}
\arguments{
\item{experimentId}{ID of experiment to which to upload.}

\item{attachmentPath}{Path to attachment.}
}
\description{
Uploads an attachment to an experiment.
}
\examples{
\dontrun{
uploadAttachment(experimentId, "/path/to/file")
}
}
