% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadFcsFile.R
\name{uploadFcsFile}
\alias{uploadFcsFile}
\title{Upload FCS file}
\usage{
uploadFcsFile(experimentId, fcsFilePath)
}
\arguments{
\item{experimentId}{ID of experiment to which to upload.}

\item{fcsFilePath}{Path to FCS file.}
}
\description{
Uploads an FCS file to an experiment.
}
\examples{
\dontrun{
uploadFcsFile(experimentId, "/path/to/file.fcs")
}
}
