% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFcsFile.R
\name{annotateFcsFile}
\alias{annotateFcsFile}
\title{Annotate FCS file}
\usage{
annotateFcsFile(experimentId, fcsFileId, annotations)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{fcsFileId}{ID of FCS file.}

\item{annotations}{List of annotations}
}
\description{
Sets the annotations on an FCS file.
}
\details{
This function overwrites any existing annotations. To add new annotations
while keeping the existing annotations, first use \code{getFcsFile} to get
the existing annotations.
}
\examples{
\dontrun{
annotations = list("annotations 1" = 1, "annotation 2" = "value")
annotateFcsFile(experimentId, fcsFileId, annotations)
}
}
