% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate}
\usage{
authenticate(username = NA, password = NA, otp = NA)
}
\arguments{
\item{username}{Your username or email address. If omitted and in an
interactive session, a prompt will be displayed.}

\item{password}{Your password. If omitted and running in RStudio or the
getPass library is installed, a prompt will be displayed.}

\item{otp}{A one-time passcode, if your account is configured with two-factor
authentication. If omitted and in an interactive session, a prompt will be
displayed. Note that this must be a string.}
}
\description{
Authenticates with the server.
}
\details{
For security, you should not store your password in plain text, and you
should ensure that your R history does not store your password. Instead,
consider setting an environment variable (see examples).
}
\examples{
\dontrun{
authenticate() # prompts for username and password (and OTP if configured)
authenticate("username") # prompts for password (and OTP if configured)

# For non-interactive sessions, use an environment variable or other means to
# avoid saving your password in plaintext.
authenticate("username", Sys.getenv("API_PASSWORD"))
}
}
