% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEllipseGate.R
\name{createEllipseGate}
\alias{createEllipseGate}
\title{Create ellipse gate}
\usage{
createEllipseGate(
  experimentId,
  xChannel,
  yChannel,
  name,
  x,
  y,
  angle,
  major,
  minor,
  label = c(x, y),
  gid = generateId(),
  parentPopulationId = NULL,
  tailoredPerFile = FALSE,
  fcsFileId = NULL,
  locked = FALSE,
  createPopulation = is.null(fcsFileId)
)
}
\arguments{
\item{experimentId}{The ID of the experiment to which to add the gate, or a
\code{byName} expression.}

\item{xChannel}{The name of the x channel to which the gate applies.}

\item{yChannel}{The name of the y channel to which the gate applies.}

\item{name}{The name of the gate.}

\item{x}{The x centerpoint of the gate.}

\item{y}{The y centerpoint of the gate.}

\item{angle}{The angle of the ellipse in radians.}

\item{major}{The major radius of the ellipse.}

\item{minor}{The minor radius of the ellipse.}

\item{label}{Position of the label. Defaults to the midpoint of the gate.}

\item{gid}{Group ID of the gate, used for tailoring. If this is not specified,
then a new Group ID will be created.}

\item{parentPopulationId}{ID of the parent population or a \code{byName}
expression. Use \code{UNGATED} for the "ungated" population. Only used if
\code{createPopulation} is \code{TRUE}.}

\item{tailoredPerFile}{Whether or not this gate is tailored per FCS file.}

\item{fcsFileId}{ID of FCS file or a \code{byName} expression, if tailored
per file. Use \code{NULL} for the global gate in a tailored gate group.}

\item{locked}{Prevents modification of the gate via the web interface.}

\item{createPopulation}{Automatically create corresponding population.
Specify \code{parentPopulationId} if set to \code{TRUE}.}
}
\description{
Creates an ellipse gate.
}
\examples{
\dontrun{
createEllipseGate(experimentId, "FSC-A", "FSC-W", "my gate", c(1, 2, 3), c(4, 5, 6))
}
}
