% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQuadrantGate.R
\name{createQuadrantGate}
\alias{createQuadrantGate}
\title{Create quadrant gate}
\usage{
createQuadrantGate(
  experimentId,
  xChannel,
  yChannel,
  name,
  x,
  y,
  labels = NULL,
  gid = generateId(),
  gids = replicate(4, generateId()),
  parentPopulationId = NULL,
  tailoredPerFile = FALSE,
  fcsFileId = NULL,
  locked = FALSE,
  createPopulations = is.null(fcsFileId)
)
}
\arguments{
\item{experimentId}{The ID of the experiment to which to add the gate, or a
\code{byName} expression.}

\item{xChannel}{The name of the x channel to which the gate applies.}

\item{yChannel}{The name of the y channel to which the gate applies.}

\item{name}{The name of the gate. Each sector is named with this parameter
and a quadrant flag (UR, UL, LL, LR).}

\item{x}{The x coordinate (after the channel's scale has been applied).}

\item{y}{The y coordinate (after the channel's scale has been applied).}

\item{labels}{Positions of the quadrant labels. A list of four length-2 vectors in
the order: UR, UL, LL, LR. These are set automatically to the plot corners.}

\item{gid}{Top-level group ID of the gate, used for tailoring. If this is not specified,
then a new Group ID will be created. For compound gates, "gid" refers to the
top-level GID. Each sector has a unique model gid and name to which
populations must refer.}

\item{gids}{Group IDs of each sector, assigned to model.gids.}

\item{parentPopulationId}{ID of the parent population or a \code{byName}
expression. Use \code{UNGATED} for the "ungated" population. Only used if
\code{createPopulation} is \code{TRUE}.}

\item{tailoredPerFile}{Whether or not this gate is tailored per FCS file.}

\item{fcsFileId}{ID of FCS file or a \code{byName} expression, if tailored
per file. Use \code{NULL} for the global gate in a tailored gate group.}

\item{locked}{Prevents modification of the gate via the web interface.}

\item{createPopulations}{Automatically create corresponding populations.
Specify \code{parentPopulationId} if set to \code{TRUE}. Because R performs
partial matching of argument names, this can also be specified as
\code{createPopulation} for consistency with the other \code{create__Gate}
functions.}
}
\description{
Creates a quadrant gate. Quadrant gates have four sectors (upper-right,
upper-left, lower-left, lower-right), each with a name and unique gid.
}
\examples{
\dontrun{
createQuadrantGate(experimentId, "FSC-A", "FSC-W", "my gate", 160000, 200000)
}
}
