% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRectangleGate.R
\name{createRectangleGate}
\alias{createRectangleGate}
\title{Create rectangle gate}
\usage{
createRectangleGate(
  experimentId,
  xChannel,
  yChannel,
  name,
  x1,
  x2,
  y1,
  y2,
  label = c(mean(c(x1, x2)), mean(c(y1, y2))),
  gid = generateId(),
  parentPopulationId = NULL,
  tailoredPerFile = FALSE,
  fcsFileId = NULL,
  locked = FALSE,
  createPopulation = is.null(fcsFileId)
)
}
\arguments{
\item{experimentId}{The ID of the experiment to which to add the gate, or a
\code{byName} expression.}

\item{xChannel}{The name of the x channel to which the gate applies.}

\item{yChannel}{The name of the y channel to which the gate applies.}

\item{name}{The name of the gate.}

\item{x1}{The first x coordinate (after the channel's scale has been applied).}

\item{x2}{The second x coordinate (after the channel's scale has been applied).}

\item{y1}{The first y coordinate (after the channel's scale has been applied).}

\item{y2}{The second y coordinate (after the channel's scale has been applied).}

\item{label}{Position of the label. Defaults to the midpoint of the gate.}

\item{gid}{Group ID of the gate, used for tailoring. If this is not specified,
then a new Group ID will be created.}

\item{parentPopulationId}{ID of the parent population or a \code{byName}
expression. Use \code{UNGATED} for the "ungated" population. Only used if
\code{createPopulation} is \code{TRUE}.}

\item{tailoredPerFile}{Whether or not this gate is tailored per FCS file.}

\item{fcsFileId}{ID of FCS file or a \code{byName} expression, if tailored
per file. Use \code{NULL} for the global gate in a tailored gate group.}

\item{locked}{Prevents modification of the gate via the web interface.}

\item{createPopulation}{Automatically create corresponding population.
Specify \code{parentPopulationId} if set to \code{TRUE}.}
}
\description{
Creates a rectangle gate.
}
\examples{
\dontrun{
createRectangleGate(experimentId, "FSC-A", "FSC-W", "my gate", 12.502, 95.102, 1020, 32021.2)
}
}
