% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteGates.R
\name{deleteGates}
\alias{deleteGates}
\title{Delete gate}
\usage{
deleteGates(experimentId, gid, gateId = NULL, exclude = NULL)
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{gid}{ID of gate family.}

\item{gateId}{ID of gate.}

\item{exclude}{Gate ID to exclude from deletion.}
}
\description{
Deletes a gate or a tailored gate family
Works for compound gates if you specify the top-level gid. Specifying the gid of a sector (i.e. one listed in model.gids) will result in no gates being deleted.
If gateID is specified, only that gate will be deleted, regardless of the other parameters specified.
}
\examples{
\dontrun{
deleteGate(experimentId, gid)
deleteGate(experimentId, gid)
}
}
