% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEllipsePoints.R
\name{fitEllipsePoints}
\alias{fitEllipsePoints}
\title{Fit ellipse points}
\usage{
fitEllipsePoints(xy, tolerance = 1e-20, max.iter = 1000)
}
\arguments{
\item{xy}{a two-column data frame containing x and y coordinates.}

\item{tolerance}{a tolerance value (default = 0.005).}

\item{max.iter}{Maximum algorithm iterations.}
}
\value{
A named list containing:
 - "covar", the 2x2 covariance matrix of the ellipse
 - "x", the x-coordinate of the ellipse center
 - "y", the y-coordinate of the ellipse center
 - "major", the length of the major axis
 - "minor", the length of the minor axis
 - "angle", the angle of rotation
}
\description{
Calculate the minimum volume enclosing ellipsoid for a set of points
}
\details{
Uses the Khachiyan Algorithm to find the the minimum volume enclosing
ellipsoid (MVEE) of a set of points. In two dimensions, this is just
the bounding ellipse (this function is limited to two dimensions).
Adapted from https://stat.ethz.ch/pipermail/r-help/2011-March/272997.html
}
