% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{Get FCS File Events}
\usage{
getEvents(experimentId, fcsFileId, populationId = NULL, scaleSetId = NULL,
  compensation = NULL, compensatedQ = FALSE, headerQ = FALSE,
  format = "FCS", destination = NULL, overwrite = FALSE,
  subsampling = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{fcsFileId}{ID of FCS file.}

\item{populationId}{Optional ID of population, for gated events.}

\item{scaleSetId}{Optional ID of scale set. If omitted and if a single
scale set exists in the experiment, that scale set will be used.}

\item{compensation}{ID of compensation or special compensation type
(\code{UNCOMPENSATED} or \code{FILE_INTERNAL}) to use for gating. This
should generally match what you used to create your gates. Not required if
\code{compensatedQ} is \code{FALSE} and exporting ungated files.}

\item{compensatedQ}{If \code{TRUE}, applies the compensation specified in
compensation to the exported events. For TSV format, the numerical values
will be the compensated values. For FCS format, the numerical values will
be unchanged, but the file header will contain the compensation as the
spill string (file-internal compensation).}

\item{headerQ}{for TSV format only: if true, file will contain a header row.}

\item{format}{One of "TSV" or "FCS".}

\item{destination}{Optional, if specified, write the file to the specified
path instead of returning it as a binary blob.}

\item{overwrite}{Optional, if a destination is specified, allows destination
file to be overwritten.}

\item{subsampling}{Optional, subsampling parameters in the form
\code{list(preSubsampleN = number, preSubsampleP = number,
postSubsampleN = number, postSubsampleP = number, seed = number)}.
Subsample-N options specify absolute subsampling values, and subsample-P options
specify a fractional subsampling value (0 to 1); specify only one pre-
and/or one post-subsample option. Specify a \code{seed} for reproducible
downsampling.}
}
\description{
Retrieves events from an FCS file or gated population as either an FCS or
TSV file.
}
\details{
Tip: FCS files are more compact than TSV files. Use that format for faster
downloads. Use a library such as flowCore to parse FCS files.
}
\examples{
\dontrun{
getEvents(experimentId, fcsFileId)
getEvents(experimentId, fcsFileId, populationId, format = "TSV")
getEvents(experimentId, fcsFileId, destination = "/path/to/output.fcs")
getEvents(experimentId, fcsFileId, populationId, subsampling = list(preSubsampleN = 5000, seed = 1.5))
}
}
