% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiment.R
\name{getExperiment}
\alias{getExperiment}
\title{Get experiment}
\usage{
getExperiment(experimentId)
}
\arguments{
\item{experimentId}{ID of experiment.}
}
\description{
Retrieves an experiment.
}
\examples{
\dontrun{
# Retrieve by ID
getExperiment(experimentId)

# Lookup by name
getExperiment(byName("my experiment"))
}
}
