% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{getExperiments}
\title{Get experiments}
\usage{
getExperiments(params = list())
}
\arguments{
\item{params}{Optional query parameters.}
}
\description{
Retrieves the list of experiments to which you have access.
}
\examples{
\dontrun{
getExperiments()
getExperiments(params = list("limit" = "5"))
}
}
