% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFcsFiles.R
\name{getFcsFiles}
\alias{getFcsFiles}
\title{Get FCS files}
\usage{
getFcsFiles(experimentId, params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{params}{Optional query parameters.}
}
\description{
Retrieves the list of FCS files in an experiment. This does not download the
FCS files themselves; this only returns information about the FCS files.
}
\examples{
\dontrun{
getFcsFiles(experimentId)
getFcsFiles(experimentId, params = list("limit" = "5"))
}
}
