% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopulations.R
\name{getPopulation}
\alias{getPopulation}
\title{Get population}
\usage{
getPopulation(experimentId, populationId, params = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{populationId}{ID of the population.}

\item{params}{Optional query parameters.}
}
\description{
Retrieves a population by ID
}
\examples{
\dontrun{
getPopulation(experimentId, populationId)
getPopulation(experimentId, populationId, params = list("fields" = "+name"))
}
}
