% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToFlowCore.R
\name{toFlowCore}
\alias{toFlowCore}
\title{Convert a gate to flowCore}
\usage{
toFlowCore(cellengineObject)
}
\arguments{
\item{cellengineObject}{The CellEngine object to be converted.}
}
\description{
Converts a CellEngine object to its flowCore analogue.
}
