% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{UNGATED}
\alias{UNGATED}
\title{Ungated.}
\format{
An object of class \code{character} of length 1.
}
\usage{
UNGATED
}
\description{
A constant representing the ungated population.
}
\details{
CellEngine uses the JSON value \code{null} to represent this in API requests
and responses. However, working with \code{NULL} is cumbersome in R. Thus,
the empty string is used instead.
}
\keyword{datasets}
