% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGate.R
\name{updateGate}
\alias{updateGate}
\title{Update gate}
\usage{
updateGate(experimentId, gateId, properties = list())
}
\arguments{
\item{experimentId}{ID of experiment.}

\item{gateId}{ID of gate or a \code{byName} expression.}

\item{properties}{Properties to set on the gate.}
}
\description{
Updates a gate.
}
\examples{
\dontrun{
updateGate(experimentId, gateId, list("name" = "new gate name"))
}
}
